Move description editor to conversation settings
